requirements.txt
aniso8601==1.2.1
astroid==1.5.3
beautifulsoup4==4.6.0
blinker==1.4
boto==2.48.0
bz2file==0.98
certifi==2017.4.17
chardet==3.0.4
click==6.7
cssselect==1.0.1
decorator==4.1.2
docutils==0.12
feedfinder2==0.0.4
feedgenerator==1.7
feedparser==5.2.1
Flask==0.12.2
Flask-RESTful==0.3.6
gensim==2.3.0
httplib2==0.10.3
idna==2.5
isort==4.2.15
itsdangerous==0.24
jieba==0.38
jieba3k==0.35.1
Jinja2==2.9.6
Keras==2.0.6
lazy-object-proxy==1.3.1
lxml==4.1.1
Markdown==2.6.5
MarkupSafe==1.0
mccabe==0.6.1
networkx==1.11
newspaper3k==0.2.5
nltk==3.2.5
numpy==1.13.1
olefile==0.44
pelican==3.6.3
Pillow==4.3.0
Pygments==2.0.2
pylint==1.7.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.2
requests-file==1.4.2
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
smart-open==1.5.3
snownlp==0.12.3
textrank4zh==0.3
Theano==0.9.0
tldextract==2.2.0
Unidecode==0.4.18
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.10
you-get==0.4.939
